split chunkviewmodel desc+value
put them into their own partial classes because I don't want Seb to have merge conflicts
