+++ WolvenKit.App/ViewModels/Shell/ChunkViewModel.cs
@@ -2198,509 +2198,7 @@ public partial class ChunkViewModel : ObservableObject, ISelectableTreeViewItemM
     #endregion
 
     #region methods
-
-    [MemberNotNull(nameof(Value))]
-    private void CalculateValue()
-    {
-        Value = "";
-
-        if (Data is null)
-        {
-            Value = "null";
-        }
-        if (PropertyType.IsAssignableTo(typeof(IRedString)) && Data is IRedString s)
-        {
-            var value = s.GetString();
-            if (!string.IsNullOrEmpty(value))
-            {
-                Value = value;
-                if (Value.StartsWith("LocKey#") && ulong.TryParse(Value[7..], out var key))
-                {
-                    Value = "";
-                }
-            }
-        }
-        else if (PropertyType.IsAssignableTo(typeof(CByteArray)) && Data is CByteArray b)
-        {
-            var ba = (byte[])b;
-            Value = string.Join(" ", ba.Select(x => $"{x:X2}"));
-        }
-        else if (PropertyType.IsAssignableTo(typeof(LocalizationString)) && Data is LocalizationString l)
-        {
-            var value = l;
-            Value = value.Value is "" or null ? "null" : value.Value;
-        }
-        else if (PropertyType.IsAssignableTo(typeof(IRedEnum)) && Data is IRedEnum e)
-        {
-            var value = e;
-            Value = value.ToEnumString();
-        }
-        else if (PropertyType.IsAssignableTo(typeof(IRedBitField)) && Data is IRedBitField f)
-        {
-            var value = f;
-            Value = value.ToBitFieldString();
-        }
-        else if (ResolvedData is CKeyValuePair kvp)
-        {
-            // If the CValuePair has a value, we'll try to resolve it
-            Value = kvp.Value switch
-            {
-                CName cname => cname.GetResolvedText() ?? "",
-                CResourceReference<ITexture> reference => reference.DepotPath.GetResolvedText() ?? "",
-                _ => kvp.Value.ToString()
-            };
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is meshMeshAppearance { ChunkMaterials: not null } appearance)
-        {
-            Value = string.Join(", ", appearance.ChunkMaterials);
-            Value = $"[{appearance.ChunkMaterials.Count}] {Value}";
-            IsValueExtrapolated = true;
-        }
-        // Material instance (mesh): "[2] - engine\materials\multilayered.mt" (show #keyValuePairs)
-        else if (ResolvedData is CMaterialInstance { BaseMaterial: { } cResourceReference } material)
-        {
-            var numMaterials = $"[{material.Values?.Count ?? 0}] - ";
-            Value = $"{numMaterials}{cResourceReference.DepotPath.GetResolvedText() ?? "none"}";
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is CResourceAsyncReference<IMaterial> materialRef)
-        {
-            Value = materialRef.DepotPath.GetResolvedText();
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is workWorkEntryId id)
-        {
-            Value = $"{id.Id}";
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is workWorkspotAnimsetEntry animsetEntry)
-        {
-            Value = $"{animsetEntry.Rig.DepotPath.GetResolvedText() ?? "none"}";
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is CMeshMaterialEntry materialDefinition)
-        {
-            Value = materialDefinition.IsLocalInstance ? "" : " (external)";
-            Value = $"{materialDefinition.Index}{Value}";
-            IsValueExtrapolated = true;
-        }
-        else if (NodeIdxInParent > -1 && ResolvedData is physicsRagdollBodyInfo &&
-                 GetRootModel().GetModelFromPath("ragdollNames")?.ResolvedData is CArray<physicsRagdollBodyNames>
-                     ragdollNames)
-        {
-            var ragdollName = ragdollNames[NodeIdxInParent];
-            Value =
-                $"{ragdollName.ParentAnimName.GetResolvedText() ?? ""} -> {ragdollName.ChildAnimName.GetResolvedText() ?? ""}";
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is animRigRetarget retarget)
-        {
-            Value = $"{retarget.SourceRig}";
-            IsValueExtrapolated = true;
-        }
-        else if (ResolvedData is redTagList list)
-        {
-            Value = $"[ {string.Join(", ", list.Tags.ToList().Select(t => t.GetResolvedText() ?? "").ToArray())} ]";
-            IsValueExtrapolated = true;
-        }
-        else if (NodeIdxInParent > -1 && Parent?.Name == "referenceTracks" &&
-                 GetRootModel().GetModelFromPath("trackNames")?.ResolvedData is CArray<CName> trackNames)
-        {
-            Value = trackNames[NodeIdxInParent].GetResolvedText();
-            IsValueExtrapolated = true;
-        }
-        //else if (PropertyType.IsAssignableTo(typeof(TweakDBID)))
-        //{
-        //    Value = (TweakDBID)Data.ToString();
-        //    //Value = Locator.Current.GetService<TweakDBService>().GetString(value);
-        //}
-        else if (PropertyType.IsAssignableTo(typeof(CBool)) && Data is CBool cb)
-        {
-            var value = cb;
-            Value = value ? "True" : "False";
-        }
-        else if (PropertyType.IsAssignableTo(typeof(CRUID)) && Data is CRUID cr)
-        {
-            var value = cr;
-            Value = ((ulong)value).ToString();
-        }
-        else if (PropertyType.IsAssignableTo(typeof(CUInt64)) && Data is CUInt64 uInt64)
-        {
-            var value = uInt64;
-            Value = value != 0 ? ((NodeRef)(ulong)value).ToString() : ((ulong)value).ToString();
-        }
-        else if (PropertyType.IsAssignableTo(typeof(gamedataLocKeyWrapper)))
-        {
-            //var value = (gamedataLocKeyWrapper)Data;
-            //Value = ((ulong)value).ToString();
-            //Value = Locator.Current.GetService<LocKeyService>().GetFemaleVariant(value);
-        }
-        else if (PropertyType.IsAssignableTo(typeof(IRedInteger)) && Data is IRedInteger i)
-        {
-            var value = i;
-
-            Value = value.ToString(CultureInfo.CurrentCulture);
-        }
-        else if (PropertyType.IsAssignableTo(typeof(FixedPoint)) && Data is FixedPoint fp)
-        {
-            var value = fp;
-            Value = ((float)value).ToString("G9");
-        }
-        else if (PropertyType.IsAssignableTo(typeof(NodeRef)) && Data is NodeRef nr)
-        {
-            var value = nr;
-            Value = value;
-        }
-        else if (PropertyType.IsAssignableTo(typeof(IRedRef)) && Data is IRedRef rr)
-        {
-            // if a path is resolved in the file, but is not yet added to the list of known hashes, add it
-            var depotPath = rr.DepotPath;
-            if (!_hashService.Contains(depotPath) && !ResourcePath.IsNullOrEmpty(depotPath))
-            {
-                _hashService.AddCustom(depotPath.GetResolvedText() ?? "");
-            }
-
-            if (depotPath.IsResolvable)
-            {
-                Value = depotPath.GetResolvedText().NotNull();
-            }
-            else
-            {
-                Value = depotPath == ResourcePath.Empty ? "null" : $"{(ulong)depotPath}{_hashService.GetGuessedExtension(depotPath)}";
-            }
-        }
-        else if (Data is IBrowsableType ibt)
-        {
-            Value = ibt.GetBrowsableValue();
-        }
-        else if (ResolvedData is animPoseLink link)
-        {
-            if (link.Node is not null)
-            {
-                var desc = GetNodeName(link.Node);
-
-                if (!string.IsNullOrEmpty(desc))
-                {
-                    Value = desc;
-                    IsValueExtrapolated = true;
-                    return;
-                }
-            }
-        }
-        
-        // factory.csv
-        else if (Parent is { Name: "compiledData" } && GetRootModel().Data is C2dArray &&
-                 Data is CArray<CString> { Count: 3 } ary)
-        {
-            IsValueExtrapolated = true;
-            Value = ary[1];
-        }
-        // i18n.json
-        else if (Data is localizationPersistenceOnScreenEntry i18n)
-        {
-            IsValueExtrapolated = true;
-            // fall back to male variant only if female variant is
-            Value = i18n.FemaleVariant;
-            if (Value == "" && i18n.MaleVariant != "")
-            {
-                Value = i18n.MaleVariant;
-            }
-        } 
-      
-        if (Value is null)
-        {
-            Value = "null";
-        }
-
-    }
-    
-    public void CalculateDescriptor()
-    {
-        Descriptor = "";
-
-        if (Data is worldNodeData sst && Tab is RDTDataViewModel dvm && dvm.Chunks[0].Data is worldStreamingSector wss && sst.NodeIndex < wss.Nodes.Count)
-        {
-            Descriptor = $"[{sst.NodeIndex}] {wss.Nodes[sst.NodeIndex].NotNull().Chunk.NotNull().DebugName}";
-            return;
-        }
-
-        if (Data is worldStreamingSectorDescriptor wssd)
-        {
-            Descriptor = (wssd.Data.DepotPath.GetResolvedText() ?? "")
-                .Replace("base\\worlds\\03_night_city\\_compiled\\default\\", "").Replace(".streamingsector", "");
-            return;
-        }
-
-        if (ResolvedData is IRedArray ary && ary.Count > 0)
-        {
-            // csv files and the like 
-            if (Parent is { Name: "compiledData" } && GetRootModel().Data is C2dArray csv)
-            {
-                var index = 0;
-                for (var i = 0; i < csv.CompiledHeaders.Count; i++)
-                {
-                    if (((string)csv.CompiledHeaders[i]).Contains("name", StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        index = i;
-                        break;
-                    }
-                }
-
-                Descriptor = $"{ary[index]}";
-                if (Descriptor != "")
-                {
-                    return;
-                }
-            }
-
-            Descriptor = $"[{ary.Count}]";
-        }
-        else if (ResolvedData is appearanceAppearancePart)
-        {
-            Descriptor = ((appearanceAppearancePart)ResolvedData).Resource.DepotPath.GetResolvedText() ?? "";
-            if (Descriptor != "")
-            {
-                return;
-            }
-        }
-        else if (ResolvedData is animAnimSetEntry)
-        {
-            var animation = ((animAnimSetEntry)ResolvedData).Animation?.GetValue();
-            Descriptor = animation?.GetProperty("Name")?.ToString() ?? "";
-            if (Descriptor != "")
-            {
-                return;
-            }
-        }
-        // animgraph - something is broken here. Why does the orange text go away? Why do I need the try/catch
-        // around the GetNodename?
-        else if (Data is CHandle<animAnimNode_Base> handle)
-        {
-            if (handle.Chunk is animAnimNode_Switch)
-            {
-                var inputNodes = handle.Chunk.GetProperty("InputNodes");
-                if (inputNodes is CArray<animPoseLink> animAry)
-                {
-                    var names = animAry
-                        .Select((animPoseLink) => GetNodeName(animPoseLink.Node) ?? "")
-                        .Where((name) => name != "")
-                        .ToArray();
-                    Descriptor = string.Join(", ", names);
-                    // Value = string.Join(", ", names);
-                    // IsValueExtrapolated = true;
-                    return;
-                }
-            }
-        }
-        else if (ResolvedData is IRedBufferPointer rbp && rbp.GetValue().Data is RedPackage pkg)
-        {
-            Descriptor = $"[{pkg.Chunks.Count}]";
-        }
-        else if (ResolvedData is IRedBufferPointer rbp2 && rbp2.GetValue().Data is CR2WList cl)
-        {
-            Descriptor = $"[{cl.Files.Count}]";
-        }
-        else if (ResolvedData is CKeyValuePair kvp)
-        {
-            Descriptor = kvp.Key;
-        }
-        else if (ResolvedData is questNodeDefinition qnd)
-        {
-            Descriptor = qnd.Id.ToString();
-            return;
-        }
-        else if (ResolvedData is scnSceneGraphNode sgn)
-        {
-            Descriptor = sgn.NodeId.Id.ToString();
-            return;
-        }
-        else if (ResolvedData is worldCompiledEffectInfo info)
-        {
-            Descriptor = string.Join(", ", info.PlacementInfos);
-            return;
-        }
-        else if (Data is TweakDBID tdb)
-        {
-            //Descriptor = Locator.Current.GetService<TweakDBService>().GetString(tdb);
-            Descriptor = tdb.GetResolvedText();
-            return;
-        }
-        else if (Data is gamedataLocKeyWrapper locKey)
-        {
-            Descriptor = ((ulong)locKey).ToString();
-            //Value = Locator.Current.GetService<LocKeyService>().GetFemaleVariant(value);
-        }
-        else if (Data is IRedString str)
-        {
-            var s = str.GetString();
-            if (s is not null && s.StartsWith("LocKey#") && ulong.TryParse(s[7..], out var locKey2))
-            {
-                Descriptor = locKey2.ToString();
-            }
-        }
-        //if (ResolvedData is CMaterialInstance && Parent is not null)
-        //{
-        //    if (Parent.Parent is not null && Parent.Parent.Parent is not null && Parent.Parent.Data is CMesh mesh)
-        //    {
-        //        Descriptor = mesh.MaterialEntries[int.Parse(Name)].Name;
-        //    }
-        //}
-        else if (Data is Vector3 v3)
-        {
-            Descriptor = $"{v3.X}, {v3.Y}, {v3.Z}";
-        }
-        else if (Data is Vector4 v4)
-        {
-            Descriptor = $"{v4.X}, {v4.Y}, {v4.Z}, {v4.W}";
-        }
-        else if (Data is Quaternion q)
-        {
-            Descriptor = $"{q.I}, {q.J}, {q.K}, {q.R}";
-        }
-
-        // For local and external materials
-        if (ResolvedData is CMaterialInstance or CResourceAsyncReference<IMaterial> && NodeIdxInParent > -1
-            && GetRootModel().GetModelFromPath("materialEntries")?.ResolvedData is CArray<CMeshMaterialEntry>
-                materialEntries && materialEntries.Count > NodeIdxInParent)
-        {
-            var isLocalMaterial = ResolvedData is CMaterialInstance;
-            var entry = materialEntries[NodeIdxInParent];
-
-            // If we immediately found the right material: use material name
-            if (entry.IsLocalInstance == isLocalMaterial && entry.Index == NodeIdxInParent)
-            {
-                Descriptor = entry.Name.GetResolvedText();
-                if (Descriptor != null)
-                {
-                    return;
-                }
-            }
-
-            // Else: Iterate to find material name
-            entry = materialEntries
-                .Where((e, idx) => e.Index == NodeIdxInParent && e.IsLocalInstance == isLocalMaterial)
-                .FirstOrDefault();
-
-            // Will return null if no entry is found
-            Descriptor = entry?.Name.GetResolvedText();
-
-            if (Descriptor != null)
-            {
-                return;
-            }
-        }
-        else if (ResolvedData is localizationPersistenceOnScreenEntry localizationPersistenceOnScreenEntry)
-        {
-            var desc = $"[{localizationPersistenceOnScreenEntry.PrimaryKey}]";
-            if (!string.IsNullOrEmpty(localizationPersistenceOnScreenEntry.SecondaryKey))
-            {
-                desc += $" {localizationPersistenceOnScreenEntry.SecondaryKey}";
-            }
-
-            Descriptor = desc;
-        }
-        else if (ResolvedData is inkTextureSlot texturesSlot)
-        {
-            var desc = texturesSlot.Texture.DepotPath.GetResolvedText();
-            if (!string.IsNullOrEmpty(desc))
-            {
-                Descriptor = desc;
-            }
-        }
-        else if (ResolvedData is entEffectDesc desc && desc.EffectName is CName name)
-        {
-            Descriptor = name;
-            return;
-        }
-        // mesh: boneTransforms (in different coordinate spaces)
-        else if (NodeIdxInParent > -1 &&
-                 (Parent?.Name == "boneTransforms" || Parent?.Name == "aPoseLS" || Parent?.Name == "aPoseMS") &&
-                 GetRootModel().GetModelFromPath("boneNames")?.ResolvedData is CArray<CName> boneNames &&
-                 boneNames.Count > NodeIdxInParent)
-        {
-            Descriptor = boneNames[NodeIdxInParent];
-        }
-        else if (ResolvedData is not null)
-        {
-            if (Data is IBrowsableDictionary ibd)
-            {
-                var pns = ibd.GetPropertyNames();
-                Descriptor = $"[{pns.Count()}]";
-            }
-            else if (Data is IList list)
-            {
-                Descriptor = $"[{list.Count}]";
-            }
-            else if (Data is Dictionary<string, object> dict)
-            {
-                Descriptor = $"[{dict.Count}]";
-            }
-        }
-        // some common "names" of classes that might be useful to display in the UI
-        var propNames = new string[]
-        {
-            "name", // default property
-            "partName", // ?
-            "slotName", // ?
-            "hudEntryName", // ?
-            "stateName", // ?
-            "characterRecordId", // tweak record children
-            "secondaryKey", // json
-            "femaleVariant", // also json
-            "maleVariant", // also json
-            "n", // ?
-            "componentName", // ?
-            "parameterName", // ?
-            "debugName", // e.g. nodes
-            "expressionString", // e.g. animMathExpressionNodes
-            "idleAnim", // .workspot handle work entry items
-            "category", // ?
-            "entryName", // ?
-            "className", // ?
-            "actorName", // ?
-            "sectorHash", // sectors
-            "propertyPath" // ?
-        };
-        if (ResolvedData is RedBaseClass irc)
-        {
-            foreach (var propName in propNames)
-            {
-                var prop = GetPropertyByRedName(irc.GetType(), propName);
-                if (prop is not null)
-                {
-                    var p = irc.GetProperty(prop.RedName.NotNull());
-                    if (p?.ToString() is null or "None")
-                    {
-                        continue;
-                    }
-                    Descriptor = p?.ToString();
-                    return;
-                }
-            }
-        }
-        else
-        {
-            foreach (var propName in propNames)
-            {
-                if (Data is not null)
-                {
-                    var prop = Data.GetType().GetProperty(propName);
-
-                    if (prop is not null)
-                    {
-                        var val = prop.GetValue(Data);
-                        if (val is not null)
-                        {
-                            Descriptor = val.ToString().NotNull();
-                        }
-
-                        return;
-                    }
-                }
-            }
-        }
-    }
+ 
 
     private string? GetNodeName(CWeakHandle<animAnimNode_Base>? linkNode)
     {
